EP-69: Applied git review comments, renamed modules to epsilonModules to fix import bug and changed imports.
